Experimental change: store the table content viewport before a refresh, and then restore it afterwards - ie suggestion (2) in http://groups.google.com/group/sequel-pro/browse_frm/thread/71541c6267fc08c1/2f4aeeaa7f15b268 .
I was initially in two minds as to whether this would be a good thing, but with r195/r196 removing the need to refresh for structural changes, this change does actually feel right to me - in fact, I now wish it didn't clear any filters/limit, and I'd like to preserve them as well.  Any objections?


git-svn-id: http://sequel-pro.googlecode.com/svn/trunk@197 eefc5464-0a4a-0410-82e6-d75a05d1654a